Pull out ECDC data fetcher into a function
